feat: 新增刪除課程資訊功能
