+++ ProjectPi/Controllers/CounselorsController.cs
@@ -18,7 +18,7 @@ using SixLabors.ImageSharp.Processing;
 
 namespace ProjectPi.Controllers
 {
-    [OpenApiTag("Counselors", Description = "諮商師操作功能")]
+    [OpenApiTag("Counselors", Description = "諮商師會員中心")]
     public class CounselorsController : ApiController
     {
         PiDbContext _db = new PiDbContext();
@@ -236,93 +236,6 @@ namespace ProjectPi.Controllers
             catch (Exception)
             {
                 return BadRequest("執照上傳失敗或未上傳");
-            }
-        }
-
-        /// <summary>
-        /// 新增/修改課程資訊
-        /// </summary>
-        /// <param name="view"></param>
-        /// <returns></returns>
-        [Route("api/courses")]
-        [JwtAuthFilter]
-        [HttpPost]
-        public IHttpActionResult PostCourses(ViewModel_C.Course view)
-        {
-            if (!ModelState.IsValid)
-                return BadRequest("欄位驗證有誤");
-
-            var counselorToken = JwtAuthFilter.GetToken(Request.Headers.Authorization.Parameter);
-            int counselorId = (int)counselorToken["Id"];
-
-            var productExist = _db.Products
-                .Where(c => c.CounselorId == counselorId && c.FieldId == view.FieldId)
-                .FirstOrDefault();
-            var hasFeature = _db.Features
-                .Where(c => c.CounselorId == counselorId && c.FieldId == view.FieldId)
-                .FirstOrDefault();
-
-            if (productExist != null) //已新增過該專業領域的課程資訊 -> 修改
-            {
-                for (int i = 0; i < view.Courses.Length; i++)
-                {
-                    var item = view.Courses[i];
-                    var findItem = _db.Products
-                        .Where(c => c.CounselorId == counselorId && c.FieldId == view.FieldId && c.Item == item.Item)
-                        .FirstOrDefault();
-
-                    findItem.Price = view.Courses[i].Price;
-                    findItem.Availability = view.Courses[i].Availability;
-                    _db.SaveChanges();
-                }
-
-                hasFeature.Feature1 = view.Features.Feature1;
-                hasFeature.Feature2 = view.Features.Feature2;
-                hasFeature.Feature3 = view.Features.Feature3;
-                hasFeature.Feature4 = view.Features.Feature4;
-                hasFeature.Feature5 = view.Features.Feature5;
-                _db.SaveChanges();
-
-                ApiResponse result = new ApiResponse { };
-                result.Success = true;
-                result.Message = "成功修改課程資訊";
-                result.Data = null;
-                return Ok(result);
-            }
-            else //從來沒有新增過該專業領域的課程資訊 -> 新增
-            {
-                for (int i = 0; i < view.Courses.Length; i++)
-                {
-                    var item = view.Courses[i];
-                    Product product = new Product();
-                    product.CounselorId = counselorId;
-                    product.FieldId = view.FieldId;
-                    product.Item = item.Item;
-                    product.Quantity = item.Quantity;
-                    product.Price = item.Price;
-                    product.Availability = item.Availability;
-
-                    _db.Products.Add(product);
-                    _db.SaveChanges();
-                }
-
-                Feature feature = new Feature();
-                feature.CounselorId = counselorId;
-                feature.FieldId = view.FieldId;
-                feature.Feature1 = view.Features.Feature1;
-                feature.Feature2 = view.Features.Feature2;
-                feature.Feature3 = view.Features.Feature3;
-                feature.Feature4 = view.Features.Feature4;
-                feature.Feature5 = view.Features.Feature5;
-
-                _db.Features.Add(feature);
-                _db.SaveChanges();
-
-                ApiResponse result = new ApiResponse { };
-                result.Success = true;
-                result.Message = "成功新增課程資訊";
-                result.Data = null;
-                return Ok(result);
             }
         }
     }
