+++ ProjectPi/ProjectPi.csproj
@@ -268,6 +268,7 @@
     <Compile Include="Controllers\AccountsController.cs" />
     <Compile Include="Controllers\AppointmentsController.cs" />
     <Compile Include="Controllers\CounselorsController.cs" />
+    <Compile Include="Controllers\CoursesController.cs" />
     <Compile Include="Controllers\HomeController.cs" />
     <Compile Include="Controllers\TestController.cs" />
     <Compile Include="Controllers\TimetablesController.cs" />
